First homework

+++ src/index.js
@@ -1,10 +1,11 @@
 const fs = require('fs');
 const path = require('path');
 
-const oldFolderName = process.argv[process.argv.length - 2];
+const oldFolderName = process.argv[2];
 const pathToFolder = path.join(__dirname, oldFolderName);
-const folderName = process.argv[process.argv.length - 1];
+const folderName = process.argv[3];
 const pathToNewFolder = path.join(__dirname, folderName);
+const deleteFolder = !!process.argv[4];
 
 const readDir = (base) => {
   const files = fs.readdirSync(base);
@@ -22,9 +23,11 @@ const readDir = (base) => {
             throw new Error(err);
           }
           const firstCharacter = item.toString().toLowerCase()[0];
+
           if (!fs.existsSync(path.join(pathToNewFolder, firstCharacter))) {
             fs.mkdirSync(path.join(pathToNewFolder, firstCharacter));
           }
+
           fs.writeFile(
             path.join(pathToNewFolder, firstCharacter, item),
             data,
@@ -33,11 +36,15 @@ const readDir = (base) => {
                 throw new Error(err);
               }
 
-              fs.unlink(newBase, (err) => {
-                if (err) {
-                  throw new Error(err);
-                }
-              });
+              if (deleteFolder) {
+                fs.unlink(newBase, (err) => {
+                  if (err) {
+                    throw new Error(err);
+                  }
+
+                  removeDir(newBase);
+                });
+              }
             }
           );
         });
@@ -46,6 +53,19 @@ const readDir = (base) => {
       }
     }
   });
+};
+
+const removeDir = (base) => {
+  const dir = path.parse(base).dir;
+  if (fs.existsSync(dir) && !fs.readdirSync(dir).length) {
+    fs.rmdir(dir, (err) => {
+      if (err) {
+        return console.log(err);
+      }
+
+      removeDir(dir);
+    });
+  }
 };
 
 if (!fs.existsSync(pathToNewFolder)) {
